added authorization to do action on bookmark

+++ src/bookmark/bookmark.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateBookmarkDto, EditBookmarkDto } from './dto';
 
@@ -39,23 +39,36 @@ export class BookmarkService {
     bookmarkId: number,
     dto: EditBookmarkDto,
   ) {
-    const bookmark = await this.prisma.bookmark.update({
+    // get bookmark by id
+    const bookmark = await this.prisma.bookmark.findUnique({
+      where: { id: bookmarkId },
+    });
+    // need to check bookmark is owned by user
+    if (!bookmark || bookmark.userId !== userId) {
+      throw new ForbiddenException('Access to resources denied');
+    }
+    return await this.prisma.bookmark.update({
       where: {
         id: bookmarkId,
-        userId,
       },
       data: {
         ...dto,
       },
     });
-    return bookmark;
   }
 
   async deleteBookmarkById(userId: number, bookmarkId: number) {
+    // get bookmark by id
+    const bookmark = await this.prisma.bookmark.findFirst({
+      where: { id: bookmarkId },
+    });
+    // need to check bookmark is owned by user
+    if (!bookmark || bookmark.userId !== userId) {
+      throw new ForbiddenException('Access to resources denied');
+    }
     await this.prisma.bookmark.delete({
       where: {
         id: bookmarkId,
-        userId,
       },
     });
   }
